Добавление моделей, фильтров

+++ src/main/java/com/example/electronicshop/entity/AbstractElectronicEntity.java
@@ -22,7 +22,7 @@ public class AbstractElectronicEntity {
      * Идентификатор
      */
     @Id
-    private BigDecimal id;
+    private Long id;
 
     /**
      * Наименование
